Refactor TeamRecord and TeamRecords
