+++ mlbstatsapi/models/teams/__init__.py
@@ -1,1 +1,2 @@
 from .team import Team
+from .attributes import TeamRecord
